+++ webknossos/Changelog.md
@@ -17,6 +17,7 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 - Renamed the  `Annotation.get_agglomerate_graph` to `Annotation.get_agglomerate_graph_data`. [#1435](https://github.com/scalableminds/webknossos-libs/pull/1435)
 - Renamed `RemoteDataset.download` argument `bbox` to `bounding_box`. `bbox` is still available, but deprecated.  [#1435](https://github.com/scalableminds/webknossos-libs/pull/1435)
 - Moved the `s3fs` dependency to an optional extra. Install with `webknossos[s3]` or `webknossos[all]`. [#1443](https://github.com/scalableminds/webknossos-libs/pull/1443)
+- The `use_bioformats` argument of `Dataset.add_layer_from_images`, `Dataset.from_images` and `RemoteDataset.from_images` is now `False` by default. `None` is not allowed anymore. [#1446](https://github.com/scalableminds/webknossos-libs/pull/1446)
 
 ### Added
 - Added `isVisible` attribute to `Skeleton.Tree` [#1442](https://github.com/scalableminds/webknossos-libs/pull/1442)